Only define ENKITS_API, if not defined.
Only define ENKITS_API, if not defined already. Allows setting as a local project setting to override. Case was experimenting with using enkiTS inside UE4 as well as outside for some code to run tasks under both. UE4 defines an API macro as part of the plugin process, so defining ENKITS_API using that made it compatible with UE4.